Create README.md
Potholer detects and displays potholes near you. It runs in the background while you do other things. 

Some things of note:
I've only tested this on a Sony Xperia S
I never put in catches to deal with no internet connectivity. It will probably crash if you open while it is not connected.
Try starting at a Severity of 7, then going up or down from there
I have yet to put in a clause for the same pothole being hit multiple times. If 1000 people hit the same hole, it'll show up as 1000 holes on the map.